+++ constant/constant.go
@@ -10,7 +10,6 @@ import (
 var (
 	ADDRESS   = ":"
 	PORT      = 8888
-	PPROF     = false
 	MEMLIMIT  = 20
 	GCPERCENT = 100
 	DataDir   string // 数据根目录
@@ -20,7 +19,6 @@ func ParseArgs() {
 	home, _ := os.UserHomeDir()
 	flag.StringVar(&ADDRESS, "address", ":", "http address")
 	flag.IntVar(&PORT, "port", 8888, "http port")
-	flag.BoolVar(&PPROF, "add-pprof", false, "add pprof")
 	flag.IntVar(&MEMLIMIT, "mem", 20, "memory limit(MB)")
 	flag.IntVar(&GCPERCENT, "gc", 100, "gc percent")
 	flag.StringVar(&DataDir, "data-dir", filepath.Join(home, ".aiapi"), "data root directory")