+++ main.go
@@ -4,7 +4,6 @@ import (
 	"io"
 	"log/slog"
 	"net/http"
-	"net/http/pprof"
 	"os"
 	"runtime/debug"
 	"time"
@@ -43,9 +42,6 @@ func main() {
 	e.HideBanner = true
 	framework.EchoInit(e)
 	ServeFrontend(e)
-	if constant.PPROF {
-		registerRoutes(e)
-	}
 	registerRuntime()
 	e.Logger.Fatal(e.StartServer(&http.Server{
 		Addr:              constant.Address(),
@@ -95,19 +91,4 @@ func initLogger() {
 func registerRuntime() {
 	debug.SetMemoryLimit(int64(constant.MEMLIMIT * 1024 * 1024))
 	debug.SetGCPercent(constant.GCPERCENT)
-}
-
-func registerRoutes(engine *echo.Echo) {
-	router := engine.Group("")
-	// 下面的路由根据要采集的数据需求注册，不用全都注册
-	router.GET("/debug/pprof", echo.WrapHandler(http.HandlerFunc(pprof.Index)))
-	router.GET("/debug/pprof/allocs", echo.WrapHandler(http.HandlerFunc(pprof.Index)))
-	router.GET("/debug/pprof/block", echo.WrapHandler(http.HandlerFunc(pprof.Index)))
-	router.GET("/debug/pprof/goroutine", echo.WrapHandler(http.HandlerFunc(pprof.Index)))
-	router.GET("/debug/pprof/heap", echo.WrapHandler(http.HandlerFunc(pprof.Index)))
-	router.GET("/debug/pprof/mutex", echo.WrapHandler(http.HandlerFunc(pprof.Index)))
-	router.GET("/debug/pprof/cmdline", echo.WrapHandler(http.HandlerFunc(pprof.Cmdline)))
-	router.GET("/debug/pprof/profile", echo.WrapHandler(http.HandlerFunc(pprof.Profile)))
-	router.GET("/debug/pprof/symbol", echo.WrapHandler(http.HandlerFunc(pprof.Symbol)))
-	router.GET("/debug/pprof/trace", echo.WrapHandler(http.HandlerFunc(pprof.Trace)))
 }